feat: leave, leaveentitlement, report and many repair

+++ database/migrations/2025_02_20_153045_create_attendances_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->foreignId('schedule_id')->nullable()->constrained('schedules')->onDelete('cascade');
 
-            $table->integer('date');
+            $table->bigInteger('date');
 
             //history
             $table->time('start_time')->nullable();
